hex2int: move to lldp_utils
This will be used by an upcoming commit adding a unit test framework.
